feat: use new logger
